improve the way AdBlock compares a URL against the block list
For context, each line in the block list is a hostname.

Previously, the hostname of a requested URL
needed to EXACTLY match a line in the block list.

Now, conceptually, each hostname in the block list
can be thought of as being prefixed by: "*."

Which is to say, that the hostname of a requested URL
will be blocked if it ends with any line in the block list.

For example:
  * the block list currently includes the line:
      yieldlab.net
  * previously:
      - the following URL was blocked:
          http://yieldlab.net/
      - the following URL was allowed:
          http://www.yieldlab.net/
  * now:
      - both URLs are blocked
  * note:
      - when allowed..
        both URLs currently return a 200 OK response,
        with an HTML body that simply contains: "OK"